Rename IDataSyncService file
